@@ -37,6 +37,9 @@ declare module 'react-navigation' {
 
   import {
     Animated,
+    Text,
+    TextInput,
+    StatusBar,
     TextStyle,
     ViewProps,
     ViewStyle,
@@ -763,6 +766,7 @@ declare module 'react-navigation' {
   }
 
   export interface NavigationNavigatorProps<O = {}, S = {}> {
+    theme?: SupportedThemes;
     detached?: boolean;
     navigation?: NavigationProp<S>;
     screenProps?: ScreenProps;
@@ -1256,7 +1260,9 @@ declare module 'react-navigation' {
   export namespace SwitchActions {
     const JUMP_TO: 'Navigation/JUMP_TO';
 
-    function jumpTo(options: NavigationJumpToActionPayload): NavigationJumpToAction;
+    function jumpTo(
+      options: NavigationJumpToActionPayload
+    ): NavigationJumpToAction;
   }
 
   /**
@@ -1521,4 +1527,40 @@ declare module 'react-navigation' {
   }
 
   export class SceneView extends React.Component {}
+
+  /**
+   * Themes
+   */
+
+  // Context
+  export type SupportedThemes = 'light' | 'dark';
+  export const ThemeContext: React.Context<SupportedThemes>;
+
+  // Hooks
+  export function useTheme(): SupportedThemes;
+
+  // Colors
+  export interface Theme {
+    header: string;
+    headerBorder: string;
+    body: string;
+    bodyBorder: string;
+    bodyContent: string;
+    bodyContentBorder: string;
+    label: string;
+  }
+  export const ThemeColors: { [k in SupportedThemes]: Theme };
+
+  // Themed components
+  interface ThemedStatusBarProps
+    extends React.ComponentProps<typeof StatusBar> {}
+  interface ThemedTextProps extends React.ComponentProps<typeof Text> {}
+  interface ThemedTextInputProps
+    extends React.ComponentProps<typeof TextInput> {}
+
+  export namespace Themed {
+    export const StatuBar: React.ComponentType<ThemedStatusBarProps>;
+    export const Text: React.ComponentType<ThemedTextProps>;
+    export const TextInput: React.ComponentType<ThemedTextInputProps>;
+  }
 }
